### 2018-07-23 #239 @wrightaprilm
submission - treestartr
This package helps users to build a reasonable starting phylogeny for phylogenetic analysis.
In particular, this package allows researchers to add tips to a tree through either a graphical
user interface, an automated system based on taxonomy, or by inputting a dataframe of taxa to
be placed.

```
Package: treestartr
Title: Generate Starting Trees For Combined Molecular, Morphological and Stratigraphic Data
Version: 0.1.0
Authors@R: person("April", "Wright", role = c("aut", "cre"),
	email = "april.wright@selu.edu")
Description: Combine a list of taxa with a phylogeny to generate a starting tree for use in
    total evidence dating analyses.
	phytools (>= 0.6-4),
	ape
    covr,
    utils
VignetteBuilder: utils

```
https://github.com/wrightaprilm/treeStartR

[e.g., "data extraction, because the package parses a scientific data file format"]

data munging, combining data sources for use with phylogenetic analysis programs (in particular RevBayes & Mr. Bayes)

As the complexity of phylogenetic models increases, it can be difficult to find starting phylogenetic trees for Bayesian MCMC that have computable likelihoods. Most programs start from a random addition tree, which, for large datasets, may not produce reasonable likelihoods. Most software allows users to input a starting tree. The starting tree is often gleaned from the literature. However, for people performing "total evidence" or "joint" phylogenetic analysis, there may not be a published tree with all the relevant taxa. This package allows users to easily create one. Fossil record taxa are commonly left off phylogenies, so I expect biologists and paleontologists working with fossils to be the chief users of this package.

Not to my knowledge.

Emailed @sckott to discuss how to know if the package is ready, and if reviewers can help with a WARNING, but did not discuss specific scope questions.

- [x ] does not violate the Terms of Service of any service it interacts with.

I have two, and I'm unsure what they mean:

Undocumented code objects:
  ‘absent_list’ ‘ant_mrca_df’ ‘ant_taxa’ ‘ant_tree’ ‘mm’ ‘mol’
  ‘mrca_df’ ‘new_tree’ ‘tax_file’ ‘tax_list’ ‘tree’
Files in the 'vignettes' directory but no files in 'inst/doc':
  ‘vignette.Rmd’, ‘vignette.pdf’

Klaus Schliep - KlausVigo
David Bapst - dwbapst
Rachel Warnock - rachelwarnock
