### 2018-07-27 #239 @sckott
thanks for your submission @wrightaprilm !

Doing editor checks right now. On trying to install I get

```r
➜  treeStartR git:(master) Rscript -e 'library(devtools); document(); install()'
Updating treestartr documentation
Loading treestartr
Error: 'ants' is not an exported object from 'namespace:treestartr'
```

can you fix? then i'll continue with editor checks
