+++ README.Rmd
@@ -48,15 +48,15 @@ library(treestartr)
 data(bears)
 ```
 
-First, we need to load a list of the total set of taxa present in the tree. The "total set" refers to any taxa that will be included in your analysis. This can be either a CSV or a TSV file. A sample list, tax_list, has been provided as part of the bears data object, but you can also generate one using the function dataf_parser. 
+First, we need to load a list of the total set of taxa present in the tree. The "total set" refers to any taxa that will be included in your analysis. This can be either a CSV or a TSV file. A sample list, bears, has been provided as part of the bears data object, but you can also generate one using the function dataf_parser. 
 
 ```{r eval=FALSE}
-tax_list <- dataf_parsr("bears_taxa.tsv")
+tax_list <- dataf_parsr("inst/extdata/bears_taxa.tsv")
 ```
 
 Next, we find out which of the taxa from our total set are not represented on the tree already. This function takes as input a tree, with or without branch lengths, but without annotations (such as 95% HPDs). It also takes the total set of taxa generated by dataf_parser:
 
-```{r}
+```{r warning=FALSE}
 absent_list <- genera_strippr(tree, tax_list)
 ```
 
@@ -64,8 +64,9 @@ absent_list <- genera_strippr(tree, tax_list)
 
 Finally, we add the tips that are not present to the tree. If there are other representatives of the same genera as an absent taxon (for example, adding an additional "Ursus" species to the example tree), those taxa will be used to place the tip. If there are multiple species of the genera, the new tip will subtend the most recent common ancestor of the tips already on the tree. If there is only one representative, the tip will subtend the parent node of that taxon.
 
-```{r}
+```{r warning=FALSE}
 new_tree <- present_tippr(tree, absent_list)
+plot(new_tree)
 ```
 
 ## Adding tips manually
@@ -79,16 +80,18 @@ new_tree <- absent_tippr(tree, absent_list)
 ## Adding tips at random
 Or, if there are no congeners, you may choose to add tips at random:
 
-```{r}
+```{r warning=FALSE}
 new_tree <- rand_absent_tippr(tree, absent_list)
+plot(new_tree)
 ```
 
 ## Adding tips via CSV
 
 Lastly, you may have a TSV file that specifies the tips to be added, and a taxon set. treeStartR will locate the MRCA of the taxon set, and add the tips subtending that node.
 
-```{r}
+```{r warning=FALSE}
 new_tree <- text_placr(tree, mrca_df)
+plot(new_tree)
 ```
 
 ## Outputting results
