+++ vignettes/vignette.Rmd
@@ -14,7 +14,6 @@ knitr::opts_chunk$set(echo = TRUE)
 ![logo](../images/logo1.png)
 
 
-
 ## treeStartR
 
 Phylogenetic trees, and particularly time-scaled phylogenetic trees, are 
@@ -27,12 +26,11 @@ allowing the user to specify the tree. Addition of taxa is usually performed bas
 The purpose of this package is to allow users to efficiently add taxa to a given tree to generate a reasonable starting tree. Functions in this package allow taxa to be added to a tree according to either their taxonomy (if other specimens from the same genus are present) in function present_tippr, at random (rand_absent_tippr), or via other user-specified groupings (text_placr). The package uses functions from 
 phytools [@phytools] and ape [@ape].
 
-
 ## Installation
 
 Install the development version directly from Github
 
-```{r}
+```{r eval=FALSE}
 library(devtools)
 install_github("wrightaprilm/treeStartR")
 ```
@@ -52,56 +50,57 @@ library(treestartr)
 data(bears)
 ```
 
-First, we need to load a list of the total set of taxa present in the tree. The "total set" refers to any taxa that will be included in your analysis. This can be either a CSV or a TSV file. A sample list, tax_list, has been provided as part of the bears data object, but you can also generate one using the function dataf_parser. 
+First, we need to load a list of the total set of taxa present in the tree. The "total set" refers to any taxa that will be included in your analysis. This can be either a CSV or a TSV file. A sample list, bears, has been provided as part of the bears data object, but you can also generate one using the function dataf_parser. 
 
 ```{r eval=FALSE}
-tax_list <- dataf_parsr("bears_taxa.tsv")
+tax_list <- dataf_parsr("inst/extdata/bears_taxa.tsv")
 ```
 
 Next, we find out which of the taxa from our total set are not represented on the tree already. This function takes as input a tree, with or without branch lengths, but without annotations (such as 95% HPDs). It also takes the total set of taxa generated by dataf_parser:
 
-```{r}
+```{r warning=FALSE}
 absent_list <- genera_strippr(tree, tax_list)
 ```
 
 ## Adding tips with congeners
 
-Finally, we add the tips that are not present to the tree. If there are other representatives of the same genera as an absent taxon (for example, adding an additional "Ursus" species to the example tree), those taxa will be used to place 
-the tip. If there are multiple species of the genera, the new tip will subtend 
-the most recent common ancestor of the tips already on the tree. If there is 
-only one representative, the tip will subtend the parent node of that taxon.
+Finally, we add the tips that are not present to the tree. If there are other representatives of the same genera as an absent taxon (for example, adding an additional "Ursus" species to the example tree), those taxa will be used to place the tip. If there are multiple species of the genera, the new tip will subtend the most recent common ancestor of the tips already on the tree. If there is only one representative, the tip will subtend the parent node of that taxon.
 
-```{r}
+```{r warning=FALSE}
 new_tree <- present_tippr(tree, absent_list)
+plot(new_tree)
 ```
 
 ## Adding tips manually
 
 We can also add the tips that have no congeners. This function will ask for input. A pop-up will be produced, showing node labels. When the program asks for input, you will tell it what node you would like the tip to subtend:
 
-```{r eval=FALSE, include=FALSE}
-new_tree <- absent_tippR(tree, absent_list)
+```{r eval=FALSE}
+new_tree <- absent_tippr(tree, absent_list)
 ```
 
 ## Adding tips at random
 Or, if there are no congeners, you may choose to add tips at random:
 
-```{r, warning=FALSE}
[…]
