### 2018-07-30 #239 @wrightaprilm
Quick question -  a user noted a (both run time and unit) test I should add for their use case, and I have added it. Should I push that now, or wait?
