### 2018-07-31 a507aab
adding runtime + unit tests for higher-order taxa
