+++ man/taxon_testr.Rd
@@ -1,0 +1,17 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/taxon_testr.R
+\name{taxon_testr}
+\alias{taxon_testr}
+\title{Utility function to check taxon_names in parsed data file}
+\usage{
+taxon_testr(tax_list)
+}
+\arguments{
+\item{dataf}{Taxon list}
+}
+\value{
+None
+}
+\description{
+Utility function to check taxon_names in parsed data file
+}
