### 2018-07-31 41822c4
adding call to higher-order tax test func
