### 2018-08-03 #239 @sckott
### Editor checks:

---

#### Editor comments

Thanks for your submission @wrightaprilm  !

Here's the output from [goodpractice](https://github.com/MangoTheCat/goodpractice/). If you haven't used `goodpractice` it's an R package that checks a number of things with another package - most of which we agree with and want authors to follow. You don't need to address these now, it's info for reviewers to use to get started.

```r
── GP treestartr ─────────
It is good practice to

    R/absent_tippR.R:15:NA
    R/absent_tippR.R:16:NA
    R/absent_tippR.R:19:NA
    R/absent_tippR.R:21:NA
    R/absent_tippR.R:22:NA
    ... and 28 more lines

  ✖ avoid long code lines, it is bad for readability. Also, many people prefer editor windows that are about 80 characters wide. Try make your lines
    shorter than 80 characters

    R/taxon_testr.R:11:1
    R/tree_df.R:20:1

  ✖ avoid sapply(), it is not type safe. It might return a vector, or a list, depending on the input data. Consider using vapply() instead.

    R/make_absent_df.R:14:23
    R/tree_df.R:24:21

  ✖ not import packages as a whole, as this can cause name clashes between the imported packages. Instead, import only the specific functions you need.
  ✖ fix this R CMD check NOTE: absent_tippr: no visible global function definition for ‘plot’ dataf_parsr: no visible global function definition for
    ‘read.csv’ test_dataf: no visible global function definition for ‘assert_that’ Undefined global functions or variables: assert_that plot read.csv Consider
    adding importFrom("graphics", "plot") importFrom("utils", "read.csv") to your NAMESPACE file.
  ✖ fix this R CMD check WARNING: Undocumented code objects: ‘absent_list’ ‘absent_tippr’ ‘dataf_parsr’ ‘genera_strippr’ ‘mrca_df’ ‘present_tippr’
    ‘rand_absent_tippr’ ‘text_placr’ ‘tree’ Undocumented data sets: ‘absent_list’ ‘mrca_df’ ‘tree’ All user-level objects in a package should have documentation
    entries. See chapter ‘Writing R documentation files’ in the ‘Writing R Extensions’ manual.
  ✖ fix this R CMD check WARNING: Data with usage in documentation object 'ants' but not in code: ants
```

Seeking reviewers now 🕐

---

Reviewers: xxx
