### 2018-09-12 #239 @dwbapst
Sorry, Scott, I sent you some emails a few weeks ago, which I later got
weird bounces back from - they were about what a review looks like. I've
got a heap of QA issues, but not sure what format I should put these into,
or what level of code analysis you want me to do?
-Dave

On Wed, Sep 12, 2018 at 2:30 PM Scott Chamberlain <notifications@github.com>
wrote:

> @dwbapst <https://github.com/dwbapst> @rachelwarnock
> <https://github.com/rachelwarnock> Will you be able to get your review in
> soon? We're about 2 weeks past due date. Thanks!
>
