### 2018-09-13 #239 @dwbapst
No problem - I'm sorry that I didn't investigate after the bounces,
but I think they came just as classes were beginning...
-Dave

On Wed, Sep 12, 2018 at 2:34 PM Scott Chamberlain
<notifications@github.com> wrote:
>
> Can you send that email again to myrmecocystus@gmail.com
>
