### 2018-09-19 #239 @rachelwarnock
Hi Scott,

I’m so sorry about the delay. I’ll try to wrap my comments & review up by
the end of next week.

Best wishes,

Rachel

On Wed, 12 Sep 2018 at 21:30, Scott Chamberlain <notifications@github.com>
wrote:

> @dwbapst <https://github.com/dwbapst> @rachelwarnock
> <https://github.com/rachelwarnock> Will you be able to get your review in
> soon? We're about 2 weeks past due date. Thanks!
>
