### 2018-09-19 #239 @dwbapst
Hi, I feel the same as Rachel. I'm new to rOpenSci reviewing and it just came at a bad time when I was teaching two graduate classes. In order to force myself to work on this, I'm going to work the draft live here in the issue thread.

	* Before getting the request, Ih ad already read the manuscript for this software, and promised I'd eventually look at the software, so this is what I was going to do anyway.

     - yeah this exists

- [ x] **Installation instructions:** for the development version of package and any non-standard dependencies in README

- [ x] **Vignette(s)** demonstrating major functionality that runs successfully locally
    - the vignette is basically identical to the top-level readme, which does seem to be demonstrating major functionality, yes

- [x ] **Function Documentation:** for all exported functions in R help
      - yes... dunno why you're asking this as R CHECK would complain about this if it weren't true

     -

>##### For packages co-submitting to JOSS
>
>- [ ] The package has an **obvious research application** according to [JOSS's definition](http://joss.theoj.org/about#submission_requirements)
>
>The package contains a `paper.md` matching [JOSS's requirements](http://joss.theoj.org/about#paper_structure) with:
>
>- [ ] **A short summary** describing the high-level functionality of the software
>- [ ] **Authors:**  A list of authors with their affiliations
>- [ ] **A statement of need** clearly stating problems the software is designed to solve and its target audience.
>- [ ] **References:** with DOIs for all those that have one (e.g. papers, datasets, software).

I strongly making a manual page named treeStartR or with that alias so that users who cannot remember a function name but can remember the package name can quickly access the documentation. This page also serves as a good starting point for users interested in the package, and you could put a lot of information from your readme there, such as directing them to the vignette.

```
absent_list <- genera_strippr(tree, bears)
```

First, the 'morphological & stratigraphic genera' ??!!
# it expects taxon names are column labeled 'taxon'

new_tree <- present_tippr(tree, absent_list)

layout(1:2)
plot(tree);plot(new_tree)

# also true of absent_tippr

#Adding tips with congeners on tree:Indarctos_punjabiensis
#Adding tip via MRCA at 25
#Adding tips with congeners on tree:Ursus_abstrusus
#Adding tip via MRCA at 30

# At what? node numbers 25 and 30?

#What does this mean??
#Warning message:

# EXAMPLE:
> tree2<-drop.tip(tree,"Indarctos_vireti")
> new_tree <- present_tippr(tree2, absent_list)
> layout(1:2)
> plot(tree);plot(new_tree)

########

new_tree <- absent_tippr(tree, absent_list)

# may want to clarify the prompt to users

new_tree <- rand_absent_tippr(tree, absent_list)

# Where would you like to place the tip?Zaragocyon_daamsi
#Error in if (where <= length(tree$tip.label) && position == 0) { :
#  missing value where TRUE/FALSE needed
#In addition: Warning messages:
#1: In bind.tree(tree, tip, where = where, position = pp) :
#  one tree has no branch lengths, they have been ignored
#2: In absent_tippr(tree, absent_list) : NAs introduced by coercion

# cannot make two species sister to each other, can only place them in a polytomy together

new_tree <- text_placr(tree, mrca_df)

# Again, this means you cannot make add a taxon as sister to another taxon.

####################

# with paleotree

library(paleotree)

# generate Genus Names
genNames<-paste0("R",
	apply(combn(c("a","b","c","d","e"),2),
		2,paste0,collapse="")
	)
# generate species names
specNames<-apply(combn(c("a","b","c","d"),3),
		2,paste0,collapse="")
#
#combine
extraSpec<-as.vector(sapply(genNames,function(x)
	paste0(x,"_",specNames)))
tipSpec<-paste0(genNames,"_regularis")
#
#lets make our hypothetical simulated tree of higher taxa
singleSpeciesTree <- rtree(10,tip.label=tipSpec)
#
# format extraSpec correctly
taxonDF<-data.frame(taxon=extraSpec)
#
library(treestartr)
absentList <- genera_strippr(singleSpeciesTree , taxonDF)

# why did it print the absent_list? This is overly noisy

    Does the code comply with general principles in the Mozilla reviewing guide?
    Does the package comply with the rOpenSci packaging guide?
    Are there improvements that could be made to the code style?
    Is there code duplication in the package that should be reduced?
    Are there user interface improvements that could be made?
    Are there performance improvements that could be made?
    Is the documentation (installation instructions/vignettes/examples/demos) clear and sufficient?
    If you have your own relevant data/problem, work through it with the package. You may find rough edges and use-cases the author didn’t think about.

    Checking the package’s logs on its continuous integration services (Travis-CI, Codecov, etc.)
    Running devtools::check() and devtools::test() on the package to find any errors that may be missed on the author’s system.
    Using the covr package to examine the extent of test coverage.
    Using the goodpractice package (goodpractice::gp()) to identify likely sources of errors and style issues. Most exceptions will need to be justified by the author in the particular context of their package.
    Using spelling::spell_check_package() (and spelling::spell_check_files("README.Rmd")) to find spelling errors.

---
