### 2018-09-21 #239 @wrightaprilm
Thanks for that, @dwbapst. Really helpful stuff here (and thankfully all addressable)! @sckott, presumably I wait for @rwarnock, in case any of her review conflicts with Dave's. But after her's comes in, it looks like in the review guidelines, I can just start work on revision without waiting on an "Editor decision" in the way of a more traditional journal?

[If I'm wrong about the above, I'm certainly happy to contribute different wording to policies.html]
