### 2018-10-01 #239 @rachelwarnock
- [X] As the reviewer I confirm that there are no conflicts of interest for me to review this work.

	- it doesn't say how to contribute but I'm not sure this is necessary

- [ ] **Automated tests:** Unit tests cover essential functions of the package.
   	- I'm not sure the tests do run; see my comment towards the end about running `devtools::check`

---

#### Comments based on the README

Love the logo!

The main functionality of the package could be stated earlier on in the README. Its not until the end of third paragraph that you learn the main purpose of the package. You could include a sentence at the beginning, saying something like, "This package produces starting trees incorporating fossils for phylogenetic analysis", before jumping into your nice intro. Or the title of the README could be "treeStartR: generating starting trees for phylogenetic analysis".

The package depends on the package phytools, but since you use the `phytools::` and `ape::` function calling format (which is great programming practice) you don't have to load the phytools library on start up for the `treestartr` functions to work (I tested to make sure).

Its useful to have the function names within text using the markdown coding format, i.e. `dataf_parser` instead of dataf_parser. That way, the functions are highlighted if you want to quickly scan the README to locate info.

I don't *think* you need to use `data(bears)` since this dataset is already loaded into the environment when you load the package. `data(ants)` returns "data set 'ants' not found", although this dataset has a help page.

With the bears dataset, the `absent_list`, `mrca_df` and `tree` components of the data are available, but not `tax_list`. Instead, the info that should be represented by this variable appears to represented by the `bears` variable. i.e. the vector output by `genera_strippr(tree, bears)` matches `absent_list`. This doesn't seem intentional.

With `new_tree <- present_tippr(tree, absent_list)`, I don't really follow the screen output from this command. Why is "Tree tip names formatted correctly" printed to screen so many times (= 33)? length(absent_list) + length(tree$tip.label) = 22 only. With a very large dataset you could end up generating a lot of unneccessary screen output. Why not just print "all tip names formatted correctly" once after checking? or just return incorrectly formatted names. This also applies to other functions that generate this output (`absent_tippr`, `rand_absent_tippr`, `get_lost`, `get_found`).

Unclear what the source of this warning message means:
In bind.tree(tree, tip, where = where, position = pp) :
  one tree has no branch lengths, they have been ignored

I suppose this has something to do with the fact that the functions are returning trees with no branch lengths. This suggests that the fossil ages are not being taken into account. Will this not be important for generating starting trees for fbd/tip-dating analyses?

The manual option is tricky for trees even as large as the example dataset. For example, on my small laptop screen, some of the branching events are too close for me to be able to distinguish the node labels (i.e. many are plotted on top of each other). Not sure if there's anything you can do about this though.

Where it says TSV file, do you mean CSV? or CSV/TSV

Perhaps it would be useful to have an example, where multiple options are used to add taxa to the tree.

#### Comments based on the help pages

`bears` - `tax\_list` not available
`ants` - has a help page but I can't seem to load the data

`absent_tippr` - the documentation say tips are added "via user input" but its not clear that this is function is interactive.

`dataf_parsr` - The function description mentions morphological and molecular data but I think this function is designed to read the file containing the stratigraphic info only.
Maybe include a url to the relevant RB tutorial? Where it describes the function argument `dataf`, I think this should be "file name" not "data frame" (which is the output). In addition, because the functions don't return time scaled trees, I'm sure why the fossil ages are required.

`genera_strippr` - no issues here.

`get_found`, `get_lost` - The distinction between these functions is a bit confusing. Could the dataframe output by this function have more useful column headers? Or perhaps these could be private internal functions.

`make_absentdf` - Perhaps this could also be a private internal function.

`present_tippr` - I would add the more comprehensive description of this function from the README here.

`rand_absent_tippr` - no issues here.

`text_placr` - maybe some additional clarification would be useful here. Am I right in thinking the `clade` column really refers to a branch/tip? And that the function then builds a clade from the tips (e.g. Ursus_arctos, Ursus_spelaeus, Ursus_americanus) from which it works out where to place the taxon (e.g. Ursus_abstrusus). At first when I looked at `mrca_df` I wasn't clear why there were multiple entries per taxon in the table.

I think it would be useful to have more complete examples in the documentation. e.g. for `rand_absent_tippr`.

```
data(bears)
absent_list <- genera_strippr(tree, tax_list)
new_tree <- rand_absent_tippr(tree, absent_list)
write.nexus(new_tree)
```
That way, if the user wanted to use this function, they would know which other function help pages to refer to get all the info necessary to generate input.

On a related note, I initially found the relationship between the package functions hard to get my head around, while just scrolling through the help pages. I produced the following for reference.

```
dataf_parsr (load a total list of taxa) ->
genera_strippr (find out which taxa are already on the tree) ->
absent_tippr (add tips manually) /
present_tippr (add tips with congeners)/
rand_absent_tippr (add tips randomly) /
text_placr (add tips using csv) ->
write.nexus
```
Perhaps you could consider including a similar flow chart or table in the README. Definitely not necessary, esp. if you included more start-to-finish examples, so its just a thought!

#### Comments based on the code

My comments on the code are extremely minor. The functions all appear to be concise and the more complicated functions are nicely commented. I can't see any obvious room for improvement in terms of efficiency or redundancy, apart from one small thing noted below.

I'm not sure each function needs to be in a separate file (i.e. related functions could potentially live in the same file) but this is down to programmer's preference and since the package isn't massive it doesn't really matter.

`genera_strippr` - why does this function print and return `absent`?

#### Other comments

Note that running `devtools::check()` produces 1 error, 4 warnings and 3 notes. Might be good to double check these issues. `devtools::spell_check()` identified one misspelling of phylogeny.

I simulated a dataset to explore the functions a bit more. Everything seems to work well but I noted the following:

- it seems to be the case that the fossil ages are not needed in `dataf` (i.e. the age column can be absent and the functions still work).
- when the fossil ages are included, I can't see that the functions use this info.
- I couldn't reproduce the errors observed with the `bears` dataset. i.e. the functions output trees with branch lengths and don't return a warning about this.
- I just noticed the functions return non-binary trees. Unsure if this is a problem.

If the functions could return time scaled starting trees that respect the fossil ages I think this would be extremely useful.

If I had to pick a favourite function it would be `text_placr`. This is super useful :)

Note, I didn't look at Dave's review before writing this, but I can see some overlap. Hope this helps!
