### 2018-10-01 #239 @sckott
@dwbapst do you have a rough estimat of hrs spent doing your review? we like to keep track for use in a long term goal of using our reviewers time most effectively
