### 2018-10-01 #239 @dwbapst
@sckott Honestly? It probably took more than 8 hours, maybe 10 hours, spread over a month? This isn't an atypical amount of time for to spend on reviewing, particularly software - what I did here is precisely I have done for reviewing papers describing new R packages at MEE. Some reviews have absolutely taken more than 10 hours, and many of my non-software reviews take longer as well.
