### 2018-10-01 #239 @dwbapst
@rachelwarnock I think that the functions can return non-binary trees is intentional, and sort of the point.
