### 2018-10-02 #239 @wrightaprilm
Thanks so much to both of you, @dwbapst and @rachelwarnock, and to you, @sckott for facilitating. I'll probably get started on response to reviews (i.e., just fixing things) Friday.
