### 2018-10-02 #239 @rachelwarnock
> @rachelwarnock I think that the functions can return non-binary trees is intentional, and sort of the point.

@dwbapst can you really specify unresolved starting trees in revbayes and beast2?
