### 2018-10-02 #239 @dwbapst
@rachelwarnock Yup, that was a dumb comment of mine. I don't know specifically what the starting tree requirements are, but you are probably right that they probably can't be non-binary.

I think I was implicitly imagining that you'd add OTUs to the initial tree, making a big informal taxon-tree sort of thing, but then for starting purposes, you'd want to randomly resolve those, and probably get some rough starting dates for the tree if for tip-dating analyses, as you talk about in your review above. I recall Nick Matzke using paleotree for that step to get starting trees for the BEAST2 analyses in the Archeopteryx paper that April and I are on. But those are steps that treestartr could help automate too.
