### 2018-10-03 #239 @rachelwarnock
@dwbapst @wrightaprilm I asked one of the BEAST2 developers (Tim Vaughan) if you can specify unresolved starting trees and he said: "Yes you can!  This will just work.  Read it in as a newick string with a polytomy.  This will be automatically converted to a binary tree with a zero edge length by beast’s newick parser, and the operators will then quickly replace the zero-length edge with one of finite-length." Maybe worth clarifying in the package documentation! I presume this is also true in RB.
