### 2018-10-03 #239 @dwbapst
@rachelwarnock Yeah, I can't find anything about starting tree requirements for RevBayes, but certainly MrBayes only requires binary trees:

http://mrbayes.sourceforge.net/Help/usertree.html

@wrightaprilm So, while I realize MrBayes is on its way out, its still fairly widely used... thus, a comment in the documentation for `treetstartr` to that effect, that MrBayes will only accept binary trees would probably reduce some confused emails from lost users.
