### 2018-10-03 #239 @wrightaprilm
Hey @dwbapst @rachelwarnock -

RB does read in polytomy trees, and it's a bit of analytical priggishness that `treestartr` returns trees with polytomies. I'll make note of it in the docs, to resolve with `multi2di` or similar if you're planning to use MB. And actually, this line of thinking is giving me an idea for something to add to the package -
exportation of RB-formatted clade constraints that include your fossils (see [here](https://github.com/revbayes/revbayes_tutorial/blob/master/RB_TotalEvidenceDating_FBD_Tutorial/scripts/model_FBDP_TEFBD.Rev)).

Again, thanks so much for all your help kicking the tires on this thing.
