### 2018-10-05 #239 @wrightaprilm
Just making a TODO-list. Ignore, unless I've misrepesented your points ;)

[ ] Making a manual page named treeStartR or with that alias so that users who cannot remember a function name but can remember the package name
[ ] absent_tippr - the documentation say tips are added "via user input" but its not clear that this is function is interactive.
[ ] ants - has a help page but I can't seem to load the data
[ ] dataf_parsr - The function description mentions morphological and molecular data but I think this function is designed to read the file containing the stratigraphic info only.
[ ] genera_strippr: This function expects a data.frame. you should be explicit about this in the help file.
[ ] genera_strippr: possibly even allow users to vary it by adding an argument like `taxonNamesVariable = "taxon"' so the default behavior stays the same.
[ ] genera_strippr: The manual page for this function refers to 'morphological & stratigraphic genera'. Its not clear to what the distinction is between these.
[ ] genera_strippr: why does this function print and return absent?
[ ] present_tippr:  very noisy.
[ ] present_tippr: clarify in docs
[ ] present_tippr, absent_tippr and text_placr.: Suppress warning.
[ ] rand_absent_tippr: clearer example
[ ] text_placr: clarify inputs
[ ] Report the number of OTUs added and the number of OTUs not added, perhaps also giving user a taxon list of the still-absent OTUs, to use with the other taxon-placing methods.
[ ] Bug: if only a single congener is on the tree, the the species gets added, but to the root node:
[ ] absent_tippr: You may want to clarify the prompt to users so they know they need to enter the number.
[ ] Defend choice: In the above code example, apparently you cannot make two species sister to each other, You can only place them in a polytomy together with a third taxon. This seems like it might be an unintentional design oversight.
[ ] Bug: text_placr: And then the taxon gets added to the root node, bizarrely.
[ ] genera_strippr: although I note that it prints the entire absentList to the console, which is overly noisy.
[ ] Man: Where it says TSV file, do you mean CSV? or CSV/TSV
[ ] ReadMe: The main functionality of the package could be stated earlier on in the README.
[ ] README: The package depends on the package phytools, but since you use the phytools:: and ape:: function calling format (which is great programming practice) you don't have to load the phytools library on start up for the treestartr functions to work (I tested to make sure).
[ ] README: ts useful to have the function names within text using the markdown coding format, i.e. dataf_parser instead of dataf_parser. That way, the functions are highlighted if you want to quickly scan the README to locate info.
[ ] Link to docs for other package functions (i.e. phytools, phangorn)
[ ] Env: I don't think you need to use data(bears) since this dataset is already loaded into the environment when you load the package. data(ants) returns "data set 'ants' not found", although this dataset has a help page.
[ ] Env: With the bears dataset, the absent_list, mrca_df and tree components of the data are available, but not tax_list. Instead, the info that should be represented by this variable appears to represented by the bears variable. i.e. the vector output by genera_strippr(tree, bears) matches absent_list. This doesn't seem intentional.
[ ] Spellcheck
[ ] get_found, get_lost, make_absentdf: Make private
[ ] Re: Ages. the immediate above: Print FBD-formatted clade constraints for RevBayes
