### 2018-10-06 591388f
supress excess warnings
