+++ R/rand_absent_tippr.R
@@ -21,7 +21,7 @@ rand_absent_tippr <- function(tree, absent_list){
 
   for (row in seq_len(nrow(lost_df))) {
     full <- as.character(lost_df[[row, "B"]])
-    message("Adding tips at random: ", full)
+    message("Adding tips at random node: ", full)
     nodel <- tree$edge[, 2]
     num <- sample(nodel, 1)
     tree <- bind.tip(tree, full, where = num)
