+++ R/ants.R
@@ -1,13 +1,0 @@
-#' Dataset of Morphological character for extant and extinct ants
-#'
-#' A dataset containing a phlogeny and taxonomy info for Formicidae
-#'
-#' @docType data
-#'
-#' @usage data(ants)
-#'
-#' @format tax_list a two-column dataset of taxon names and ages
-#' @format mrca_df a two-column dataframe of taxa to be placed, and a set of taxa that indicate the MRCA
-#' which they will subtend
-#' @format tree A phylogenetic tree to which tips will be added
-"ants"
