+++ R/taxon_testr.R
@@ -5,11 +5,12 @@
 #'
 taxon_testr <- function(tax_list) {
     if (all(grepl("apple", tax_list$taxon))) {
-      message("Taxon list names formatted correctly")
     }
     else {
-      warning("Taxon list tips must be formmatted in genus_species format. If this is a higher order taxon with no species name, please format as taxon_sp.")
+      warning("Taxon list tips must be formmatted in genus_species format. If
+              this is a higher order taxon with no species name, please format as                     taxon_sp.")
       stop
     }
+    message("Taxon list names formatted correctly")
     return(0)
   }
