+++ R/bears.R
@@ -1,14 +1,0 @@
-#' Dataset of Morphological character for extant and extinct bears
-#'
-#' A dataset containing 18 taxa and 62 characters for bears
-#'
-#' @docType data
-#'
-#' @usage data(bears)
-#'
-#' @format tax_list a two-column dataset of taxon names and ages
-#' @format mrca_df a two-column dataframe of taxa to be placed, and a set of taxa that indicate the MRCA
-#' which they will subtend
-#' @format tree A phylogenetic tree to which tips will be added
-#' @source \url{https://github.com/revbayes/revbayes_tutorial/tree/master/RB_TotalEvidenceDating_FBD_Tutorial/data}
-"bears"
