### 2018-10-06 e7978ac
add functionality to write out RB clade constraints and subtrees with taxa added
