### 2018-10-08 133ac16
last pkg mods in response to reviewer notes
