+++ R/treestartr.R
@@ -1,0 +1,10 @@
+#' treestartr: A package for generating reasonable starting trees for phylogenetic analysis.
+#'
+#' treestartr provides multiple ways to add tipts to a starting tree
+#'
+#' @section treestartr functions:
+#' absent_tippr, echo_rb, echo_subtree, present_tippr, rand_absent_tippr, text_placr
+#'
+#' @docType package
+#' @name treestartr
+NULL
