+++ R/data.R
@@ -17,19 +17,3 @@
 
 "bears"
 
-#' ants
-#'
-#' A dataset containing a phlogeny and taxonomy info for Formicidae
-#' @docType data
-#' @keywords datasets
-#'
-#'
-#' @usage data(ants)
-#' @format A collection of objects representing function inputs:
-#'\describe{
-#'     \item{tax_list}{a two-column dataset of taxon names and ages}
-#'     \item{mrca_df}{a two-column dataframe of taxa to be placed, and a set of taxa that indicate the MRCA which they will subtend}
-#'     \item{tree}{A phylogenetic tree to which tips will be added}
-#' }
-
-"ants"
