+++ R/rand_absent_tippr.R
@@ -4,9 +4,8 @@
 #' @param tree Starting tree; object of type phylo
 #' @param absent_list Vector of taxa in the total dataset that are not on the
 #' tree
-#' @param echo_subtrees Boolean; Print newick subtree with missing taxa added to screen
-#' @param echo_revbayes Boolean; Print newick subtree with missing taxa added to screen, formatted for RevBayes fossilized birth-death analysis
-#' Default FALSE.
+#' @param echo_subtrees Boolean; Print newick subtree with missing taxa added to screen. Default FALSE.
+#' @param echo_revbayes Boolean; Print clade constraints with missing taxa added to screen, formatted for RevBayes fossilized birth-death analysis. Default FALSE.
 #' @return tree Phylo object containing the starting tree,
 #'          and all tips that were added.
 #' @examples
@@ -36,17 +35,12 @@ rand_absent_tippr <- function(tree, absent_list, echo_subtrees = NULL, echo_revb
     num <- sample(nodel, 1)
     tree <- suppressWarnings(bind.tip(tree, full, where = num))
     if (!is.null(echo_revbayes)){
-      parent <- getParent(tree, num)
-      sub_list <- ape::extract.clade(tree, parent)
-      quote_vec <-paste0('"', c(sub_list$tip.label, full), '"')
-      q_vec <-paste0(quote_vec[-length(quote_vec)], ',')
-      q_final <- append(q_vec, tail(quote_vec, n=1))
-      cat("clade(", q_final, ")", '\n')
+      q_final <- echo_rb(tree, mrca_list, full)
+      cat("clade(", q_final, ")\n")
     }
     if (!is.null(echo_subtrees)){
-      parent <- getParent(tree, num)
-      sub_list <- ape::extract.clade(tree, parent)
-      cat("Subtree:", ape::write.tree(sub_list), '\n')
+      e_t <- echo_subtree(tree, mrca_list, tip)
+      cat("Subtree: ", e_t, "\n")
     }
   }
   return(tree)
