+++ R/get_lost.R
@@ -6,8 +6,7 @@
 #' @return not_found_df Dataframe objects expressing the tips that are not
 #'         on the tree, and don"t have congeners
 #' @examples
-#' no_congeners <- get_lost(absent_list, tree)
-#' @export
+#' no_congeners <- treestartr:::get_lost(absent_list, tree)
 
 get_lost <- function(absent_list, tree){
   absent_df <- make_absentdf(absent_list)
