+++ R/taxon_testr.R
@@ -8,7 +8,7 @@ taxon_testr <- function(tax_list) {
     }
     else {
       warning("Taxon list tips must be formmatted in genus_species format. If
-              this is a higher order taxon with no species name, please format as                     taxon_sp.")
+              this is a higher order taxon with no species name, please format               as taxon_sp.")
       stop
     }
     message("Taxon list names formatted correctly")
