+++ NAMESPACE
@@ -2,10 +2,9 @@
 
 export(absent_tippr)
 export(dataf_parsr)
+export(echo_rb)
+export(echo_subtree)
 export(genera_strippr)
-export(get_found)
-export(get_lost)
-export(make_absentdf)
 export(present_tippr)
 export(rand_absent_tippr)
 export(text_placr)
