+++ R/taxon_testr.R
@@ -1,14 +1,14 @@
 #' Utility function to check taxon_names in parsed data file
 #'
-#' @param dataf Taxon list
+#' @param tax_list List of taxa parsed from datafile.
 #' @return None
 #'
 taxon_testr <- function(tax_list) {
     if (all(grepl("_", tax_list$taxon))) {
     }
     else {
-      warning("Taxon list tips must be formmatted in genus_species format. If
-              this is a higher order taxon with no species name, please format               as taxon_sp.")
+      warning("Taxon list tips must be formmatted in genus_species format.
+              If this is a higher order taxon with no species name, please                   format as taxon_sp.")
       stop
     }
     message("Taxon list names formatted correctly")
