+++ R/echo_rb.R
@@ -7,6 +7,7 @@
 #' @examples
 #'  mrca_list <- c("Ursus_arctos", "Ursus_spelaeus", "Ursus_americanus")
 #' clade_constraint <- echo_rb(tree, mrca_list, "Ursus_abstrusus")
+#' @importFrom("tail")
 #' @export
 #'
 echo_rb <- function(tree, mrca_list, tip){
