+++ R/dataframe_loadR.R
@@ -1,9 +1,7 @@
 #' Detect file format of taxon file
 #' @description Detect file format of the total taxon list (molecular, morphological, and stratigraphic) and maximum age of fossil, see RevBayes total
 #' -evidence fossil file for an example of this.
-#' @param dataf A data frame with one column containing the taxon name of tips in
-#' phylogenetic tree, and one column indicating the maximum age of the tip. If
-#' the tip is extant, use 0.0 as the age. Can be CSV or TSV.
+#' @param dataf A data frame with one column containing the taxon name of tips in phylogenetic tree. Optionally, a second column can indicate the maximum age of the tip. If the tip is extant, use 0.0 as the age. Can be CSV or TSV.
 #' @return tax_list Dataframe containing the total set of tips on the tree
 #' @examples
 #' \dontrun{taxa_df <- dataf_parsr(dataf)}
