### 2018-10-08 #239 @rachelwarnock
This all sounds great to me!

> On 8 Oct 2018, at 22:24, April Wright <notifications@github.com> wrote:
>
> I've made the vast majority of the straight-forward changes discussed above in as asked. There are several that bear more discussion, though.
>
> Taxa being added to the root node: This was a bug resulting from getting the parent node of a vector, as opposed to a single taxon. It is now fixed.
> Necessity of age information in the taxa file: Rachel noted correctly that the ages are never used. This is somewhat of a bias on my part - I wanted it to be possible to just take your RevBayes taxon file, and automatically generate a starting tree from it. I've now amended the behavior of the parser to allow you to do so, but so that it only requires a column of taxa (i.e., ages are optional). RevBayes (and BEAST I believe is the same) doesn't actually parse the dates off a starting tree; it gets them from the ages file and the fossil interval file, so decorating the tree with them wouldn't be overly helpful.
> Polytomies: I added a section to the README to make it obvious that you'll need to resolve polytomies. Since we're not estimating placements from data, I pretty strongly feel that polytomies are the way to export the tree, and I think users should be forced to confront the decision to resolve them if that's what they'd like to do.
> Formatted clade constraints: I added two functions echo_subtree and echo_rb. Echo subtree simply calls ape's print subtree function to print the subtree to which the tip was added, including the tip. echo_rb created RevBayes formatted clade constraints. I added two tests for these functions.
> Ants data: I just took this out. It's for the eventual MEE paper, and I think I would like it to be in a separate repo.
>
> I have a couple CRAN check warnings that are both stubborn and cryptic. Can I ask about them?
>
> Thanks so much to all three of you for your help with this; I can already see the package is much improved.
>
