### 2018-10-09 #239 @dwbapst
Hi April, I'm afraid I encountered a number of issue, particularly some functions which broke (and one function that broke if package `ape` was loaded).

Here's my R script I was working from with commented out output for you to look at:

```
devtools::install_github("wrightaprilm/treeStartR")

library(treestartr)
packageVersion("treeStartR")

# > packageVersion("treeStartR")
# [1] ‘0.1.0’

# to build manual:
# devtools:::build_manual("d://dave//workspace//treeStartR")

data(bears)

#########################################################

absent_list <- genera_strippr(tree, bears)
absent_list
# works

# if  tax_list isn't a data.frame, it silently fails (returns no error)
	# but returns NA as the output
# Here's two examples:
bears2 <- as.vector(bears[,1]) # class character
absent_list2 <- genera_strippr(tree, bears2)
absent_list2
#
tax_list<-bears[,1] # factor
absent_list4 <- genera_strippr(tree, tax_list)
absent_list4

# example output
# > absent_list4 <- genera_strippr(tree, tax_list)
# > absent_list4
# [1] NA

# if the variable isn't named 'taxon' it can't find the taxon names, returns error
bears3 <- bears
names(bears3)[1] <- "taxa"
absent_list3 <- genera_strippr(tree, bears3)
absent_list3

# > absent_list3 <- genera_strippr(tree, bears3)
# Error in `[.data.frame`(tax_list, "taxon") : undefined columns selected
# > absent_list3
# Error: object 'absent_list3' not found

# this requirement needs to be specified in the documentation file

#####################################################
new_tree <- present_tippr(tree, absent_list)
layout(1:2)
plot(tree);plot(new_tree)
# works

# weird error if only a single congener is on the tree:
tree2<-ape::drop.tip(tree,"Indarctos_vireti")
new_tree <- present_tippr(tree2, absent_list)

# output:
# >  new_tree <- present_tippr(tree2, absent_list)
# Tree tip names formatted correctly
# [1] "Indarctos_arctoides"
# Error in getParent(tree, loc) : object 'loc' not found

# COULD NOT TRY FOLLOWING AGAIN:
# cannot make two species sister to each other,
	# can only place them in a polytomy together

#####################################

new_tree <- absent_tippr(tree, absent_list)

# this function no longer works:
# > new_tree <- absent_tippr(tree, absent_list)
# Error: 'plot' is not an exported object from 'namespace:ape'

# it looks like its looking for ape::plot.phylo but there's
	# possibly a reference to ape::plot (which doesn't exist
# plot.phylo is what we call an S3 method - we can call a
	# generic function, but R checks if there
# is a more specific function for that class first

# COULD NOT TRY FOLLOWING AGAIN:
# not immediately clear that its asking for me to type in the
	# number of the node, rather than interactively click on the plot
	# (as in some ape and phytools functions)
# may want to clarify the prompt to users

# COULD NOT TRY FOLLOWING AGAIN:
# Where would you like to place the tip?Zaragocyon_daamsi
#Error in if (where <= length(tree$tip.label) && position == 0) { :
#  missing value where TRUE/FALSE needed
#In addition: Warning messages:
#1: In bind.tree(tree, tip, where = where, position = pp) :
#  one tree has no branch lengths, they have been ignored
#2: In absent_tippr(tree, absent_list) : NAs introduced by coercion

###############################

new_tree <- rand_absent_tippr(tree, absent_list)
plot(new_tree)
# works, but...

# fatal error if run after loading package ape
#
# > library(ape)
# > new_tree <- rand_absent_tippr(tree, absent_list)
# Adding tips at random node: Ailurarctos_lufengensis
# Adding tips at random node: Kretzoiarctos_beatrix
# Error in if (tree$edge.length[which(tree$edge[, 2] == where)] <= 1e-12) { :
#  missing value where TRUE/FALSE needed

################################################
new_tree <- text_placr(tree, mrca_df)

# get this sort of output:
# > new_tree <- text_placr(tree, mrca_df)
# Placing tip Kretzoiarctos_beatrix
# via relatives Indarctos_arctoidesIndarctos_vireti
#  at node 25
# Placing tip Ursus_abstrusus
# via relatives Ursus_arctosUrsus_spelaeusUrsus_americanus
#  at node 30

# printing the node numbers for usr possibly isn't important (or maybe it is)
# may want to paste in spaces or commas between taxon names for readability

# test if only a single relative is listed:
mrca_df2 <- mrca_df[-1,]
new_tree <- text_placr(tree, mrca_df)
# works fine!
```

I hope this helps.
