+++ R/taxon_testr.R
@@ -1,16 +1,16 @@
 #' Utility function to check taxon_names in parsed data file
 #'
-#' @param tax_list List of taxa parsed from datafile.
+#' @param tax_frame Dataframe of taxa parsed from datafile.
 #' @return None
 #'
-taxon_testr <- function(tax_list) {
-    if (all(grepl("_", tax_list$taxon))) {
+taxon_testr <- function(tax_frame) {
+    if (all(grepl("_", tax_frame$taxon))) {
     }
     else {
-      warning("Taxon list tips must be formmatted in genus_species format.
+      warning("Taxon dataframe tips must be formmatted in genus_species format.
               If this is a higher order taxon with no species name, please                   format as taxon_sp.")
       stop
     }
-    message("Taxon list names formatted correctly")
+    message("Taxon dataframe names formatted correctly")
     return(0)
   }
