+++ README.Rmd
@@ -45,13 +45,13 @@ data(bears)
 First, we need to load a list of the total set of taxa present in the tree. The "total set" refers to any taxa that will be included in your analysis. This can be either a CSV or a TSV file. A sample list, bears, has been provided as part of the bears data object, but you can also generate one using the function dataf_parser. Note that if you have higher order taxa, for example, if you are placing a fossil that you have identified to family level, but not species level, these should be included as "family_spp." treeStartR relies on the underscore to separate between the higher order and lower order taxa. Use consistent formatting between the tree and taxon list - i.e, do not call a taxon "Ge_spp" on the tree and "Genus_spp" in the taxon list.
 
 ```{r}
-tax_list <- dataf_parsr("inst/extdata/bears_taxa.tsv")
+tax_frame <- dataf_parsr("inst/extdata/bears_taxa.tsv")
 ```
 
 Next, we find out which of the taxa from our total set are not represented on the tree already. This function takes as input a tree, with or without branch lengths, but without annotations (such as 95% HPDs). It also takes the total set of taxa generated by `dataf_parser`:
 
 ```{r warning=FALSE}
-absent_list <- genera_strippr(tree, tax_list)
+absent_list <- genera_strippr(tree, tax_frame)
 ```
 
 ## Adding tips with congeners
@@ -66,7 +66,7 @@ plot(new_tree)
 After we have added our tips, we can check how many tips remain to be added, and which they are:
 
 ```{r warning=FALSE}
-genera_strippr(new_tree, tax_list)
+genera_strippr(new_tree, tax_frame)
 ```
 
 ## Adding tips manually
