+++ man/dataf_parsR.Rd
@@ -7,17 +7,15 @@
 dataf_parsr(dataf)
 }
 \arguments{
-\item{dataf}{A data frame with one column containing the taxon name of tips in
-phylogenetic tree, and one column indicating the maximum age of the tip. If
-the tip is extant, use 0.0 as the age. Can be CSV or TSV.}
+\item{dataf}{A data frame with one column containing the taxon name of tips in phylogenetic tree. Optionally, a second column can indicate the maximum age of the tip. If the tip is extant, use 0.0 as the age. Can be CSV or TSV.}
 }
 \value{
-tax_list Dataframe containing the total set of tips on the tree
+tax_frame Dataframe containing the total set of tips on the tree
 }
 \description{
 Detect file format of the total taxon list (molecular, morphological, and stratigraphic) and maximum age of fossil, see RevBayes total
 -evidence fossil file for an example of this.
 }
 \examples{
-\dontrun{taxa_df <- dataf_parsr(dataf)}
+\dontrun{tax_frame <- dataf_parsr(dataf)}
 }
