### 2018-10-10 #239 @wrightaprilm
**if  tax_list isn't a data.frame, it silently fails:** Fixed - I hadn't considered the situation where someone might not use the dataframe loader function, but might be working separately with the taxa and then call treestartr. So now the `genera_strippr` function checks for dataframe, tries to coerce one it it isn't, and if that isn't possible, returns an error.
**weird error if only a single congener is on the tree::** This is a bug -  I changed a variable name in one spot, but not another.
**Error: 'plot' is not an exported object from 'namespace:ape:** I tried to solve a CRAN check error and created a namespace collision. It's fixed now.
**fatal error if run after loading package ape:** I was hilariously mystified by this, but adding in `phytools::` before `bind.tip` solves it (and I should have done anyway).

I forgot to note this in my last entry: I didn't add having the tips that _weren't_ placed echo to the screen after placing tips. That can be recalculated by calling the `genera_strippr`, and I felt like some of the functions were already getting a bit noisy, especially if you do something like print RevBayes clade constraints.

So what I'm really trying to say is I'll probably take you up on CRAN check assistance once the toothpick comes out clean on the rest of it.

Thanks for having a second look. It's so easy to get caught up in how **I** would use this package and not consider how another person might integrate it into a workflow, particularly in conjunction with more complex data processing or other packages.
