### 2018-10-10 #239 @dwbapst
Yeah, need to be explicit that you're importing `ape::plot.phylo` and `phytools::bind.tip` I think maybe ape has its own `bind.tip` with slightly different behavior.

Its absolutely not a requirement from your perspective, but I am curious if there is a specific reasoning why `ape` isn't a depends - although there is general advice to import packages rather than loading the whole package,most phylogenetics packages in R depend on ape as it is usual for users to be doing phylogenetics-things that require the S3 methods for class `phylo` in `ape`. Are you envisioning a use case where users are unlikely to have `ape` loaded?

I'll try to rerun the code later, after teaching.
