### 2018-10-11 #239 @dwbapst
Hi April, A lot more works now, but I think there is some oversight about how species are added whether to tips or lower nodes, and that sort of behavior needs to be specified in the documentation, cause I can't figure out how intentional or not this behavior is. Particularly see the worked example at the end of the following, and I'll post the plotted result as well.

```
devtools::install_github("wrightaprilm/treeStartR")

library(treestartr)
packageVersion("treeStartR")

# > packageVersion("treeStartR")
# [1] ‘0.1.0’

# to build manual:
# devtools:::build_manual("d://dave//workspace//treeStartR")

data(bears)

#########################################################

absent_list <- genera_strippr(tree, bears)
absent_list
# works

# if  tax_list isn't a data.frame, it works
# Here's two examples:
bears2 <- as.vector(bears[,1]) # class character
absent_list2 <- genera_strippr(tree, bears2)
absent_list2
# another example
tax_list<-bears[,1] # factor
absent_list4 <- genera_strippr(tree, tax_list)
absent_list4

# if the variable isn't named 'taxon' it can't find
	# the taxon names, returns error
bears3 <- bears
names(bears3)[1] <- "taxa"
absent_list3 <- genera_strippr(tree, bears3)
absent_list3

# > absent_list3 <- genera_strippr(tree, bears3)
# Error in `[.data.frame`(tax_list, "taxon") : undefined columns selected

# Some options (choose one, or combine several):
# Use the one column provided if it a one-column data table
	# or use the first column as a default?
	# but issue an informative warning() saying you are doing so
# Issue an informative error message with stop()
	# that it can't find the taxon names as expect
	# saying the function expects a certain variable name
# Specify this requirement in the documentation file

#####################################################
new_tree <- present_tippr(tree, absent_list)
layout(1:2)
plot(tree);plot(new_tree)
# works

# drop Indarctos, see if it can place two species as sister
	# to each other, or if it makes a polytomy:
tree2<-ape::drop.tip(tree,"Indarctos_vireti")
new_tree <- present_tippr(tree2, absent_list)
layout(1:2)
plot(tree);plot(new_tree)

# it places them in a polytomy instead, making the genus paraphyletic
# is this intentional behavior?

#####################################

new_tree <- absent_tippr(tree, absent_list)
# nice clear interaction :D

# Although not labeled, we can also give tip ID numbers
	# and thus add taxa as sister taxa on the tree
# Is this intentional behavior? It isn't described in documentation.
# If this is intentional behavior it should be described in documentation.

###############################

new_tree <- rand_absent_tippr(tree, absent_list)
plot(new_tree)
# works even with ape loaded

################################################
new_tree <- text_placr(tree, mrca_df)

# test if only a single relative is listed:
mrca_df2 <- mrca_df[-1,]
new_tree <- text_placr(tree, mrca_df2)
plot(new_tree)

# Hmm, adds it to a polytomy with both Indarctos species
	# Is this intentional behavior?

####################

# constructed example

set.seed(444)
# generate Genus Names
genNames<-paste0("R",
	apply(combn(c("a","b","c","d","e"),2),
		2,paste0,collapse="")
	)
# generate species names
specNames<-apply(combn(c("a","b","c","d"),3),
		2,paste0,collapse="")
#
#combine
extraSpec<-as.vector(sapply(genNames,function(x)
	paste0(x,"_",specNames)))
tipSpec<-paste0(genNames,"_regularis")
#
#lets make our hypothetical simulated tree of higher taxa
singleSpeciesTree <- rtree(10,tip.label=tipSpec)
#
# format extraSpec correctly
taxonDF<-data.frame(taxon=extraSpec)
#
library(treestartr)
absentList <- genera_strippr(singleSpeciesTree , taxonDF)
absentList

# just first three
newTree <- present_tippr(singleSpeciesTree , absentList[1:3])
# works

# all
newTree <- present_tippr(singleSpeciesTree , absentList)
# works

plot.phylo(newTree, no.margin=TRUE)

# its collapsed all sister genera into each other
	# this really doesn't look intentional!
```

![image](https://user-images.githubusercontent.com/1561903/46816161-f910db80-cd41-11e8-92db-c7b8de72a0b8.png)
