### 2018-10-12 9cec194
fixing an odd behavior noted by @dwbapst
