+++ NAMESPACE
@@ -5,6 +5,7 @@ export(dataf_parsr)
 export(echo_rb)
 export(echo_subtree)
 export(genera_strippr)
+export(get_lost)
 export(present_tippr)
 export(rand_absent_tippr)
 export(text_placr)
