### 2018-10-12 #239 @wrightaprilm
Thanks, again, for all these thorough comments. I really do appreciate it.

**# Error in `[.data.frame`(tax_list, "taxon") : undefined columns selected:** Bug caused by inappropriate curly brace closure.

**# Although not labeled, we can also give tip ID numbers and thus add taxa as sister taxa on the tree:** Added this to the visual, and suppressed warnings for once sister taxa are added (which gives a polytomy warning).

**it places them in a polytomy instead, making the genus paraphyletic # is this intentional behavior?**
It was not ... and I see how it doesn't make sense. I've changed this so that if there is one congener, the tip is added as sister.

**text_placr polytomy:** I genuinely would like to hear opinions on this - I like this behavior, but you seem confused by it. Is there a different behavior that would be preferable?

**Placements breaking up monophyletic groups**: I can see why this one is undesireable. This is fixed by allowing tips to be added as sister, an regenerating the mrca_list afterwards, such that any subsequent tips will be placed into the now-monophyletic grouping. The above now returns the below, which I think is the behavior we want:

![rplot](https://user-images.githubusercontent.com/1715446/46882476-760c8580-ce14-11e8-960d-88ea720050d8.png)
