### 2018-10-12 #239 @dwbapst
Hi April,

I think labeling the tips makes it clear that it is acceptable for users to add taxa as sisters to the tips. I just was pushing on the documentation issue because there weren't tip labels, so it wasn't clear if the behavior was intended or not. (Unintentional behaviors may be relied upon by others, but be unstable because you might not know someone else depends on that functionality, so its best to be clear). I think its clear now, though.

I still have a tiny issue with how `genera_strippr` picks what column to use from a data.frame, which I think you could solve just by being more explicit/detailed in the documentation for that function.

Overall, I'd say I don't have really any issue with the package now.

```
devtools::install_github("wrightaprilm/treeStartR")

library(treestartr)
packageVersion("treeStartR")

# > packageVersion("treeStartR")
# [1] ‘0.1.0’

# to build manual:
# devtools:::build_manual("d://dave//workspace//treeStartR")

data(bears)

#########################################################

absent_list <- genera_strippr(tree, bears)
absent_list
# works

# if  tax_list isn't a data.frame, it works
# Here's two examples:
bears2 <- as.vector(bears[,1]) # class character
absent_list2 <- genera_strippr(tree, bears2)
absent_list2
# another example
tax_list<-bears[,1] # factor
absent_list4 <- genera_strippr(tree, tax_list)
absent_list4

# the first row is always used if a data.frame is handed over
bears3 <- bears[,2:1]
#names(bears3)[1] <- "taxa"
absent_list3 <- genera_strippr(tree, bears3)
absent_list3

# > absent_list3 <- genera_strippr(tree, bears3)
#> absent_list3
# [1] "0"    "0"    "0"    "0"    "0"    "0"    "0"    "0"    "5"    "7.7"
#[11] "0.5"  "14"   "9.6"  "5.1"  "8.7"  "11.7" "36.6" "16.2" "14.4" "4.3"
#[21] "0.06" "21.9"

# I really do think you need to be explicit in the documentation that
# its the first column, or a column with a particular label, or whatever
# the rule is, or say so in a warning() or a message() when

# From before:
#
# Some options (choose one, or combine several):
# Use the one column provided if it a one-column data table
	# or use the first column as a default?
	# but issue an informative warning() saying you are doing so
# Issue an informative error message with stop()
	# that it can't find the taxon names as expect
	# saying the function expects a certain variable name
# Specify this requirement in the documentation file

#####################################################
new_tree <- present_tippr(tree, absent_list)
layout(1:2)
plot(tree);plot(new_tree)
# works

# drop 1 Indarctos sp, places other sp
	# as sister to the one remaining congeneric
# works!
tree2<-ape::drop.tip(tree,"Indarctos_vireti")
new_tree <- present_tippr(tree2, absent_list)
layout(1:2)
plot(tree, no.margin=TRUE);plot(new_tree,no.margin=TRUE)

#####################################

new_tree <- absent_tippr(tree, absent_list)
# nice clear interaction :D

# we can also give tip ID numbers
	# and thus add taxa as sister taxa on the tree
# Documentation doesn't rule out this behavior,
	# labeling makes it clear this is probably okay
# Cool

###############################

new_tree <- rand_absent_tippr(tree, absent_list)
plot(new_tree)
# works even with ape loaded

################################################
new_tree <- text_placr(tree, mrca_df)

# test if only a single relative is listed:
mrca_df2 <- mrca_df[-1,]
mrca_df2
new_tree <- text_placr(tree, mrca_df2)
plot(new_tree)

# K_beastrix added as a sister properly

####################

# constructed example

set.seed(444)
# generate Genus Names
genNames<-paste0("R",
	apply(combn(c("a","b","c","d","e"),2),
		2,paste0,collapse="")
	)
# generate species names
specNames<-apply(combn(c("a","b","c","d"),3),
		2,paste0,collapse="")
#
#combine
extraSpec<-as.vector(sapply(genNames,function(x)
	paste0(x,"_",specNames)))
tipSpec<-paste0(genNames,"_regularis")
#
#lets make our hypothetical simulated tree of higher taxa
singleSpeciesTree <- rtree(10,tip.label=tipSpec)
#
# format extraSpec correctly
taxonDF<-data.frame(taxon=extraSpec)
#
library(treestartr)
absentList <- genera_strippr(singleSpeciesTree , taxonDF)
absentList

# just first three
newTree <- present_tippr(singleSpeciesTree , absentList[1:3])
# works

# all
newTree <- present_tippr(singleSpeciesTree , absentList)
# works

plot.phylo(newTree, no.margin=TRUE)

# its collapsed all sister genera into each other
	# this really doesn't look intentional!
```
