### 2018-10-12 67179a8
update docs and error message
