### 2018-10-16 #239 @wrightaprilm
If we're all happy with the major package functions, is it OK if I ask some CRAN check questions?

- I'm not sure what this means: absent_tippr: no visible global function definition for ‘plot’
- Or this: Undocumented code objects: ‘absent_list’ ‘mrca_df’
- Also this: Files in the 'vignettes' directory but no files in 'inst/doc': ‘vignette.Rmd’, ‘vignette.pdf’

I feel very silly about this, because this part seems like it should be easy.
