### 2018-10-16 #239 @dwbapst
No, don't feel silly - `R CHECK --as-cran` is often one of the hardest parts because of the opaqueness of the messages.

- You need to specify what package you are importing plot from. Is that base plot in `graphics` or `ape`? You likely need to import `plot` from `graphics` - its a relatively new requirement.

- You have exported objects in the code, or arguments for exported functions that have no documentation assigned to them. You need to search for `absent_list` and `mrca_df`.

- Hmm. I would try deleting your `inst` directory. To be honest, I'm not sure why you have one to begin with, and the vignette in it is out of date, I think.
