### 2018-10-16 #239 @sckott
agree, don't feel silly @wrightaprilm !  The cran check stuff is tough and no fun for anyone.

I think you should add `inst/docs` to `.Rbuildignore`. `inst/doc` is a formal directory that is used when building the package, so we don't want to ignore that of course.

I think a problem might be in the line in your DESCRIPTION file where you have  `VignetteBuilder: utils` - I've not seen `utils` in there before. Maybe it is valid, I don't know, but usually the entry there is `knitr`

In the vignette, you have the line

```r
tax_list <- dataf_parsr("inst/extdata/bears_taxa.tsv")
```

Instead of the path for that directly, i'd use

```r
tax_list <- dataf_parsr(system.file("extdata/bears_taxa.tsv", package = "treestartr"))
```

Also, that chunk has `eval=FALSE` which means `tax_list` is never created, which means the next chunk then can't find `tax_list` and the vignette build fails
