### 2018-10-19 #239 @wrightaprilm
I'm still a little fuzzy on what it means for an exported object to have attached documentation. For example, `mrca_df` is defined as a parameter to `text_placr`, and I see it in the manual.
