### 2018-10-23 #239 @dwbapst
As an addendum, it’s usually good practice to name objects in an example
data file something different than the exact name of function arguments so
we know which one R is complaining about.

Cheers,
-Dave

On Fri, Oct 19, 2018 at 2:38 PM April Wright <notifications@github.com>
wrote:

> I'm still a little fuzzy on what it means for an exported object to have
> attached documentation. For example, mrca_df is defined as a parameter to
> text_placr, and I see it in the manual.
>
