### 2018-10-24 06d542f
last few things for CRAN check
