+++ README.Rmd
@@ -45,7 +45,7 @@ data(bears)
 First, we need to load a list of the total set of taxa present in the tree. The "total set" refers to any taxa that will be included in your analysis. This can be either a CSV or a TSV file. A sample list, bears, has been provided as part of the bears data object, but you can also generate one using the function dataf_parser. Note that if you have higher order taxa, for example, if you are placing a fossil that you have identified to family level, but not species level, these should be included as "family_spp." treeStartR relies on the underscore to separate between the higher order and lower order taxa. Use consistent formatting between the tree and taxon list - i.e, do not call a taxon "Ge_spp" on the tree and "Genus_spp" in the taxon list.
 
 ```{r}
-tax_frame <- dataf_parsr("inst/extdata/bears_taxa.tsv")
+tax_frame <- dataf_parsr(system.file("extdata/bears_taxa.tsv", package = "treestartr"))
 ```
 
 Next, we find out which of the taxa from our total set are not represented on the tree already. This function takes as input a tree, with or without branch lengths, but without annotations (such as 95% HPDs). It also takes the total set of taxa generated by `dataf_parser`:
@@ -111,7 +111,7 @@ ape::multi2di(new_tree)
 
 ## Subtrees and Clade Constraints
 
-All tip addition functions have the capability to write out [RevBayes-formatted](https://revbayes.github.io/tutorials/fbd/#Exercise-ModelFBD) clade constraint strings. The `echo_revbayes` argument allows this.
+All tip addition functions, except `absent_tippr` have the capability to write out [RevBayes-formatted](https://revbayes.github.io/tutorials/fbd/#Exercise-ModelFBD) clade constraint strings. The `echo_revbayes` argument allows this.
 
 ```{r warning=FALSE}
 text_placr(tree, mrca_df, echo_revbayes = TRUE)
@@ -128,7 +128,7 @@ rand_absent_tippr(tree, absent_list, echo_subtrees = TRUE)
 
 The final tree can be output using standard functions in ape, such as the  `write.nexus()` [function](https://www.rdocumentation.org/packages/ape/versions/5.2/topics/write.tree):
 
-```{r}
+```{r eval=FALSE}
 ape::write.nexus(new_tree, file = "data/export_example.tre")
 ```
 
