+++ NAMESPACE
@@ -5,8 +5,11 @@ export(dataf_parsr)
 export(echo_rb)
 export(echo_subtree)
 export(genera_strippr)
-export(get_lost)
 export(present_tippr)
 export(rand_absent_tippr)
 export(text_placr)
 import(phytools)
+importFrom(ape,getMRCA)
+importFrom(graphics,plot)
+importFrom(utils,read.csv)
+importFrom(utils,tail)
