### 2018-10-24 #239 @sckott
Nice work @wrightaprilm I see a few more things:

- add these to your `.Rbuildignore` file: README.Rmd, README.html, bibliography.bib, deprecated, docs, images, paper, refs.bib
- There are some undocumented objects still:

```
checking for missing documentation entries (1s)
   Undocumented code objects:
     ‘absent_tippr’ ‘genera_strippr’ ‘present_tippr
```

There's some remaining issues from my comment above https://github.com/ropensci/onboarding/issues/239#issuecomment-430371029 Did you see it yet?
