+++ DESCRIPTION
@@ -17,7 +17,8 @@ Suggests:
     knitr,
     testthat,
     covr
-VignetteBuilder: utils
+VignetteBuilder:
+    knitr
 Encoding: UTF-8
 LazyData: true
 RoxygenNote: 6.0.1
