### 2018-10-26 932b035
fixes to vignette build
