### 2018-10-26 79009e6
clean up extraneous doc folder
