### 2018-10-26 #239 @wrightaprilm
- Realized I should have been using a `inst/doc` folder all along.
- Had copied the vignettbuilder from another package, now changed to `knitr`, but this seems to generate another error:

```
Files named as vignettes but with no recognized vignette engine:
   ‘vignettes/vignette.Rmd’
(Is a VignetteBuilder field missing?).
```
- I don't see the undocumented code objects issue - is there a flag I need to be using to get that?
