### 2018-10-26 #239 @sckott
I'd recommend running check with cran set to `TRUE`, e.g., I run from the command line like

```
Rscript -e 'devtools::check(cran = TRUE)'
```

In your vignette you're using rmarkdown - you need to add that pkg to your Suggests
