### 2018-11-02 141a490
updating documentation and moving vignettes
