+++ deprecated/parse_morphology.Rd
@@ -1,17 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/parse_morphological.R
-\name{parse_morphology}
-\alias{parse_morphology}
-\title{Load morpohological matrices to get taxon lists}
-\usage{
-parse_morphology(morphDat)
-}
-\arguments{
-\item{morphDat}{A Nexus file containing morphological matrix}
-}
-\value{
-morph_ns A vector of names present in either the matrix
-}
-\description{
-Load morpohological matrices to get taxon lists
-}
