+++ DESCRIPTION
@@ -16,9 +16,9 @@ Depends:
 Suggests:
     knitr,
     testthat,
+    rmarkdown,
     covr
-VignetteBuilder:
-    knitr
+VignetteBuilder: knitr
 Encoding: UTF-8
 LazyData: true
-RoxygenNote: 6.0.1
+RoxygenNote: 6.1.0
