### 2018-11-02 eac939b
adding small patch to rand_abs
