### 2018-11-02 #239 @wrightaprilm
Hi Scott, I don't get that same error you were getting when I run that check. My output is below.

I did edit some documentation, and fix a DESCRIPTION error, and change a behavior after a student ran into a problem. Undergrads are the best testers besides maybe @dwbapst ;).

Let me know if you still get the same thing. I don't have anything that is modified and not pushed in my workspace.

Thanks!
