### 2018-11-08 #239 @sckott
@wrightaprilm looking good now!

I see that `phytools` and `ape` have moved into `Depends`. I see @dwbapst comment above about this. it seems like you can just import the s3 methods, like `@importFrom ape plot.phylo`, etc. or is there some reason that doesn't work. Depends surely fixes this, but is it really necessary?
