### 2018-11-08 #239 @dwbapst
Hmm. I should explain the background of my comment. So, I don't think
there's anything non-preferable about a phylogenetics package in R
depending on `ape` - in my experience, it is rare that any
phylogenetics-loaded package would be loaded without ape, and users
might often need to use the utilities in ape (and possibly, phytools)
while using a phylogenetics package. Of course, this shouldn't be a
crutch for laziness with regards to specifying functions from other
packages, though: even if you use Depends, it is still preferable to
specify ape::plot.phylo in internal code, for example.

I understand this perspective is in contrast to current best-practices
recommendations from Hadley and some other authorities, their
recommendation being that nothing nothing should ever be depended.
Obviously, a package author should make the best choice for their
package, and their programming style!
-Dave
On Wed, Nov 7, 2018 at 8:06 PM Scott Chamberlain
<notifications@github.com> wrote:
>
> @wrightaprilm looking good now!
>
> I see that phytools and ape have moved into Depends. I see @dwbapst comment above about this. it seems like you can just import the s3 methods, like @importFrom ape plot.phylo, etc. or is there some reason that doesn't work. Depends surely fixes this, but is it really necessary?
>
