+++ R/echo_rb.R
@@ -1,7 +1,7 @@
 #' Print RevBayes-formatted clade constriants.
 #' @description Once a tip has been added to the tree, print the subtree to which the tip was added. The output of this function is printed for use as a RevBayes clade constraint object.
 #' @param tree Tree to which the tip will be added; object of type phylo
-#' @param mrca_list Vector of taxa. The added tip will subtend the MRCA of these taxa.
+#' @param mrca_list Vector of taxa. The added tip will subtend the MRCA of these taxa
 #' @param tip Taxon which will be added to the tree.
 #' @return vector A comma-separated list of all the taxa in the subtree to which the tip was added.
 #' @examples
