+++ tests/testthat/test-echo-sub.R
@@ -1,3 +1,5 @@
+library(ape)
+
 test_that("Test echo_subtree(tree, mrca_df)", {
   tree <- read.nexus("testdata/simple.tre")
   mrca_df <- read.csv("testdata/mrca_df.tsv", sep = "\t")
