### 2018-11-08 #239 @sckott
Given the arguments, I can go either way, so if you think Depends is easier go for it. or Imports.

---

@dwbapst and @rachelwarnock are you both happy with the changes @wrightaprilm has made ?
