### 2018-11-14 #239 @dwbapst
Yes, I'm happy. Apologies, I was away at a week-long conference at the time
of your message, and I am only digging myself out now from the blizzard of
emails. @rachelwarnock might have been similarly delayed by the same
conference, as she was also attending, and has intercontinental travel to
boot.

On Wed, Nov 14, 2018 at 11:19 AM Scott Chamberlain <notifications@github.com>
wrote:

> @dwbapst <https://github.com/dwbapst> and @rachelwarnock
> <https://github.com/rachelwarnock> are you both happy with the changes
> @wrightaprilm <https://github.com/wrightaprilm> has made ?
>
