### 2018-11-15 #239 @rachelwarnock
Hi @Scott Chamberlain <scott@ropensci.org> - apologies for this delay! I'm
very happy with all the changes @wrightaprilm
<https://github.com/wrightaprilm> made!

On Wed, 14 Nov 2018 at 21:20, Scott Chamberlain <notifications@github.com>
wrote:

> thanks @dwbapst <https://github.com/dwbapst>
>
