### 2018-11-16 #239 @wrightaprilm
Hurray! I'll take care of the TODO over the next couple days.

I would definitely be interested in a blog post - I found this process really helpful and well-organized. I think there's a lot of value to the community even for people who aren't "developers", or whatever, and I'd like to express that in some form.
