### 2018-11-16 #239 @stefaniebutland
Hi @wrightaprilm.
I'm thrilled to hear you're interested in contributing a blog post.
> I think there's a lot of value to the community even for people who aren't "developers", or whatever, and I'd like to express that in some form.

That sentiment reflects part of our mission "Building capacity of software users and developers and fostering a sense of pride in their work" so will be very welcome in your post!

[This link](https://ropensci.org/tags/onboarding/) will give you many examples of blog posts by authors of onboarded packages so you can get an idea of the style and length you prefer.

Here are some technical and editorial guidelines for contributing a blog post: https://github.com/ropensci/roweb2#contributing-a-blog-post. We ask that you submit your draft post via pull request a week before the planned publication date so we can give you some feedback.
I have 2018-12-11 available for publication. Would submission by 2018-12-04 work for you?

Happy to answer any questions.
