### 2018-11-19 #239 @wrightaprilm
Real quick question - planning on putting some work in on this this afternoon. I had a Code of Conduct statement - Am I now covered under y'all's code of conduct? Should I edit my statement to reflect that ROpen Sci would now be in a reporting structure?

Thanks! I've never done something like this before.
