### 2018-11-19 #239 @sckott
We do have a COC for our events and any communications in our forums/slack/etc., BUT for repos, each repo should have their own COC file - which then also shows up on the right hand side of an issue when someone opens an issue. I'd suggest changing the file name to `CODE_OF_CONDUCT.md` instead of `CONDUCT.md` as I think github is looking for the former to link to
