### 2018-11-19 #239 @wrightaprilm
Sorry to belabor the point - I have this sort of thing on my mind a bit lately. If someone were to feel that I was the source of unwelcoming behavior on the repository, would they be able to report it to y'all? I found the CoC [site](https://ropensci.org/coc/), but I didn't see an email on it for someone to make a report at.

Thx!
