### 2018-11-19 9fc982e
modifying code of conduct files and updating readme
