+++ README.md
@@ -1,17 +1,10 @@
----
-title: "README.Rmd"
-output: html_document
----
-
-```{r setup, include=FALSE}
-knitr::opts_chunk$set(echo = TRUE)
-```
-![logo](images/logo1.png)
 [![Build Status](https://travis-ci.org/wrightaprilm/treeStartR.svg?branch=master)](https://travis-ci.org/wrightaprilm/treeStartR)
 [![codecov](https://codecov.io/gh/wrightaprilm/treeStartR/branch/master/graph/badge.svg)](https://codecov.io/gh/wrightaprilm/treeStartR)
-
+[![](https://badges.ropensci.org/239_status.svg)](https://github.com/ropensci/onboarding/issues/239)
 
 ## treeStartR
+
+treeStartR is an R package for auomated generation of starting trees for total-evidence phylogenetic analyses. 
 
 Phylogenetic trees, and particularly time-scaled phylogenetic trees, are 
 increasingly estimated using parameter-rich models of evolution [@catmodel; @wright2016] and models incorporating macroevoloutionary processes [@fbd]. Finding a starting tree with a computable likelihood to perform a Bayesian MCMC under these complex models can be a challenge, particularly when when estimation involves many
@@ -20,24 +13,16 @@ taxa, large datasets, and missing data.
 Different phylogenetic estimation packages allow users to find starting trees in different ways, such as estimating a tree under parsimony (@raxml8) or neighbor-joining [@beast2], randomly adding taxa to the tree [@beast2; @raxml8], or 
 allowing the user to specify the tree. Addition of taxa is usually performed based on data - that is, using an algorithm (such as parsimony or neighbor-joining) to add the tips to the tree. However, in analyses of the fossil record, specimens may be included for which there are no molecular or morphological characters available, but the taxonomy of the specimen is known via expert opinion. This is the case with many specimens harvested via repositories such as the Paleobiology Database. 
 
-The purpose of this package is to allow users to efficiently add taxa to a given tree to generate a reasonable starting tree. Functions in this package allow taxa to be added to a tree according to either their taxonomy (if other specimens from the same genus are present) in function present_tippr, at random (rand_absent_tippr), or via other user-specified groupings (text_placr). The package uses functions from 
+The purpose of this package is to allow users to efficiently add taxa to a given tree to generate a reasonable starting tree. Functions in this package allow taxa to be added to a tree according to either their taxonomy (if other specimens from the same genus are present) in function `present_tippr`, at random (`rand_absent_tippr`), or via other user-specified groupings (`text_placr`). The package uses functions from 
 phytools [@phytools] and ape [@ape].
 
 ## Installation
 
 Install the development version directly from Github
 
-```{r}
+```{r eval=FALSE}
 library(devtools)
 install_github("wrightaprilm/treeStartR")
-```
-
-## Dependencies
-
-treeStartR depends on phytools.
-
-```{r}
-library(phytools)
 ```
 
 ## Usage Example
@@ -47,29 +32,36 @@ library(treestartr)
 data(bears)
 ```
 
-First, we need to load a list of the total set of taxa present in the tree. The "total set" refers to any taxa that will be included in your analysis. This can be either a CSV or a TSV file. A sample list, tax_list, has been provided as part of the bears data object, but you can also generate one using the function dataf_parser. 
-
-```{r eval=FALSE}
-tax_list <- dataf_parsr(path/to/tax_file)
-```
-
[…]
