### 2018-11-20 e9cfe2d
transfer ownership, delete fluff, note higher-order enforcement of CoC
