### 2018-11-20 82e9651
new badge locations
