+++ README.Rmd
@@ -12,8 +12,8 @@ bibliography: bibliography.bib
 knitr::opts_chunk$set(echo = TRUE)
 ```
 ![logo](images/logo1.png)
-[![Build Status](https://travis-ci.org/wrightaprilm/treeStartR.svg?branch=master)](https://travis-ci.org/wrightaprilm/treeStartR)
-[![codecov](https://codecov.io/gh/wrightaprilm/treeStartR/branch/master/graph/badge.svg)](https://codecov.io/gh/wrightaprilm/treeStartR)
+[![Build Status](https://travis-ci.org/ropensci/treeStartR.svg?branch=master)](https://travis-ci.org/wrightaprilm/treeStartR)
+[![codecov](https://codecov.io/gh/ropensci/treeStartR/branch/master/graph/badge.svg)](https://codecov.io/gh/ropensci/treeStartR)
 [![](https://badges.ropensci.org/239_status.svg)](https://github.com/ropensci/onboarding/issues/239)
 
 ## treeStartR
